+++ test_settings.py
@@ -19,24 +19,45 @@ CACHES = {
     }
 }
 
-# Default Django middleware
+TEMPLATES = [
+    {
+        'BACKEND': 'django.template.backends.django.DjangoTemplates',
+        'APP_DIRS': True,
+    },
+]
+
 MIDDLEWARE = [
     'django.middleware.security.SecurityMiddleware',
     'django.contrib.sessions.middleware.SessionMiddleware',
     'django.middleware.common.CommonMiddleware',
+    'djangae.contrib.common.middleware.RequestStorageMiddleware',
     'django.middleware.csrf.CsrfViewMiddleware',
     'django.contrib.messages.middleware.MessageMiddleware',
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
+    'djangae.contrib.googleauth.middleware.AuthenticationMiddleware',
     'djangae.tasks.middleware.task_environment_middleware',
 ]
 
 INSTALLED_APPS = (
-    'django.contrib.sessions',
-    'gcloudc',
     'djangae',
+    'django.contrib.contenttypes',
+    'django.contrib.sessions',
+    'django.contrib.auth',
+    'djangae.contrib.googleauth',
+    'gcloudc',
     'djangae.tasks',
     'djangae.contrib.search',
 )
+
+AUTHENTICATION_BACKENDS = [
+    'djangae.contrib.googleauth.backends.iap.IAPBackend',
+    'djangae.contrib.googleauth.backends.oauth2.OAuthBackend',
+]
+
+AUTH_USER_MODEL = "googleauth.User"
+
+GOOGLEAUTH_CLIENT_ID = "test"
+GOOGLEAUTH_CLIENT_SECRET = "test"
 
 DATABASES = {
     'default': {
@@ -44,6 +65,9 @@ DATABASES = {
         'INDEXES_FILE': os.path.join(os.path.abspath(os.path.dirname(__file__)), "djangaeidx.yaml"),
         "PROJECT": "test",
         "NAMESPACE": "ns1",  # Use a non-default namespace to catch edge cases where we forget
+        "OPTIONS": {
+            "BULK_BATCH_SIZE": 25
+        }
     }
 }
 
@@ -70,4 +94,7 @@ ROOT_URLCONF = __name__
 
 urlpatterns = [
     path('tasks/', include('djangae.tasks.urls')),
+    path('_ah/', include('djangae.urls')),
 ]
+
+os.environ['OAUTHLIB_INSECURE_TRANSPORT'] = '1'